requirements.txt
crewai>=0.130.0,<1.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
streamlit>=1.40.0,<2.0.0
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
tavily-python>=0.5.0,<1.0.0
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
markdown>=3.7,<4.0.0
